% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{common}
\alias{common}
\title{Common models}
\format{A dataset with variables make, model, n (total number of models) and
  years (total number of model-years).}
\usage{
common
}
\description{
These models have at least 10 years worth of data. This dataset is
suitable for a left-join designed to restrict the number of observations
to a more manageable level.
}
\examples{
if (require("dplyr")) {
vehicles \%>\% semi_join(common)
}
}
\keyword{datasets}

